Fully removing old dim. Adding tests for gqwak.
